requirements.txt
asgiref==3.8.1
boto3==1.35.14
botocore==1.35.14
Django==5.1.1
django-environ==0.11.2
django-storages==1.14.4
jmespath==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
djangorestframework~=3.15.2
minio~=7.2.8